fix(twofa): using ref instead normal var
