+++ src/features/auth/components/TwoFACodeInput.vue
@@ -1,7 +1,9 @@
 <script setup lang="ts">
-import { ref, onMounted, nextTick, watch} from 'vue'
+import { ref, onMounted, nextTick, watch } from 'vue'
 
-const emit = defineEmits<(e: 'complete', code: string) => void>()
+const emit = defineEmits<{
+  (event: 'complete', code: string): void
+}>()
 
 const props = defineProps<{
   inputMode?: 'numeric' | 'text' | 'tel' | 'decimal' | 'none',
@@ -10,31 +12,29 @@ const props = defineProps<{
   separator?: number[]
 }>()
 
-
 const code = ref<string[]>(Array(props.length ?? 6).fill(''))
 const inputs = ref<HTMLInputElement[]>([])
-const timer = ref(59)
 const shake = ref(false)
-
-let interval: ReturnType<typeof setInterval>
 let lastCode = ''
+
+const reset = () => {
+  code.value = Array(props.length ?? 6).fill('')
+  nextTick(() => inputs.value[0]?.focus())
+}
+
+defineExpose({
+  triggerShake: () => {
+    shake.value = true
+    setTimeout(() => {
+      shake.value = false
+    }, 600)
+  },
+  reset
+})
 
 onMounted(() => {
   nextTick(() => inputs.value[0]?.focus())
-  startTimer()
 })
-
-const startTimer = () => {
-  interval = setInterval(() => {
-    if (timer.value > 0) {
-      timer.value--
-    } else {
-      clearInterval(interval)
-    }
-  }, 1000)
-}
-
-
 
 watch(code, (newVal) => {
   const current = newVal.join('')
@@ -52,12 +52,35 @@ const handleInput = (index: number, event: Event) => {
     value = value.replace(/\D/g, '')
   }
 
-  value = value.slice(0, 1)
-  code.value[index] = value
+  if (!value) return
 
-  if (value && index < code.value.length - 1) {
-    inputs.value[index + 1]?.focus()
+  const chars = value.slice(0, code.value.length - index).split('')
+  for (let i = 0; i < chars.length; i++) {
+    code.value[index + i] = chars[i]
   }
+
+  const nextIndex = index + chars.length
+  if (nextIndex < code.value.length) {
+    inputs.value[nextIndex]?.focus()
+  } else {
+    inputs.value[code.value.length - 1]?.blur()
+  }
+}
+
+const handlePaste = (event: ClipboardEvent) => {
+  event.preventDefault()
+  let pasted = event.clipboardData?.getData('text') ?? ''
+  pasted = pasted.trim()
+  if (props.inputMode === 'numeric') pasted = pasted.replace(/\D/g, '')
+  const chars = pasted.slice(0, code.value.length).split('')
+  code.value = [...chars, ...Array(code.value.length - chars.length).fill('')]
+  nextTick(() => {
+    if (inputs.value[chars.length]) {
+      inputs.value[chars.length].focus()
+    } else {
+      inputs.value[code.value.length - 1]?.blur()
+    }
+  })
 }
 
 const handleKeydown = (index: number, event: KeyboardEvent) => {
@@ -68,32 +91,21 @@ const handleKeydown = (index: number, event: KeyboardEvent) => {
   }
 }
 
-defineExpose({
-  triggerShake: () => {
-    shake.value = true
-    setTimeout(() => {
-      shake.value = false
-    }, 600)
-  }
-})
-
-
 const getCharAt = (index: number) =>
   props.placeholder?.charAt(index) || ''
-
 </script>
 
 <template>
-  <div class="flex flex-col items-center gap-4">
+  <div class="flex flex-col items-center gap-4" @paste="handlePaste">
     <div class="flex justify-center gap-2" :class="{ 'animate-shake': shake }">
       <template v-for="i in props.length ?? 6" :key="i">
         <input
+          v-model="code[i - 1]"
           type="text"
           :inputmode="props.inputMode ?? 'numeric'"
           maxlength="1"
           autocomplete="off"
           class="w-12 h-12 text-center text-xl border rounded-lg font-mono bg-white outline-none transition focus:ring-2 focus:ring-primary"
-          v-model="code[i - 1]"
           @input="handleInput(i - 1, $event)"
           @keydown="handleKeydown(i - 1, $event)"
           :placeholder="getCharAt(i - 1)"
@@ -102,19 +114,11 @@ const getCharAt = (index: number) =>
         <span
           v-if="props.separator?.includes(i)"
           class="text-xl font-bold text-neutral-dark self-center select-none"
-        >
-          -
-        </span>
+        >-</span>
       </template>
-    </div>
-
-    <div class="text-sm text-neutral-dark tracking-wide">
-      Code expirant dans :
-      <span class="font-semibold tabular-nums">{{ String(timer).padStart(2, '0') }}s</span>
     </div>
   </div>
 </template>
-
 
 <style scoped>
 @keyframes shake {
@@ -122,7 +126,6 @@ const getCharAt = (index: number) =>
   20%, 60% { transform: translateX(-5px); }
   40%, 80% { transform: translateX(5px); }
 }
-
 .animate-shake {
   animation: shake 0.4s ease;
 }
